Fix logout bug

+++ src/helper_functions/logOut.js
@@ -2,11 +2,13 @@ import axios from "axios";
 import Config from "./Config";
 
 export default function logOut(user, setUser, history) {
-  setUser(null);
-  localStorage.clear();
+  if (user === null) return;
   const config = new Config(user.token);
   axios
     .post("http://localhost:4000/logout", {}, config)
     .catch((e) => alert(e))
     .finally(() => history.push("/login"));
+  setUser(null);
+  localStorage.clear();
 }
+
